+++ app/src/main/java/com/example/findu/domain/model/my/MyInterestData.kt
@@ -3,14 +3,14 @@ package com.example.findu.domain.model.my
 data class MyInterestData(
     val interestAnimals: List<InterestAnimal>,
     val isLast: Boolean,
-    val lastInterestProtectId: Int,
-    val lastInterestReportId: Int
+    val lastInterestProtectId: Long,
+    val lastInterestReportId: Long
 ) {
     data class InterestAnimal(
-        val animalId: Int,
+        val animalId: Long,
         val date: String,
         val interest: Boolean,
-        val interestId: Int,
+        val interestId: Long,
         val isProtectingAnimal: Boolean,
         val location: String,
         val tag: String,
